backend/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
python-dotenv==1.0.1
pymongo==4.3.3
Werkzeug==2.1.2